requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
altair=4.2.0=pypi_0
appnope=0.1.3=pypi_0
argon2-cffi=21.3.0=pypi_0
argon2-cffi-bindings=21.2.0=pypi_0
asttokens=2.0.5=pypi_0
attrs=21.4.0=pypi_0
backcall=0.2.0=pypi_0
backports-zoneinfo=0.2.1=pypi_0
beautifulsoup4=4.11.1=pypi_0
bleach=5.0.1=pypi_0
blinker=1.4=pypi_0
ca-certificates=2022.4.26=hecd8cb5_0
cachetools=5.2.0=pypi_0
certifi=2022.6.15=py38hecd8cb5_0
cffi=1.15.1=pypi_0
charset-normalizer=2.1.0=pypi_0
click=8.1.3=pypi_0
commonmark=0.9.1=pypi_0
cycler=0.11.0=pypi_0
debugpy=1.6.0=pypi_0
decorator=5.1.1=pypi_0
defusedxml=0.7.1=pypi_0
entrypoints=0.4=pypi_0
executing=0.8.3=pypi_0
fastjsonschema=2.15.3=pypi_0
fonttools=4.33.3=pypi_0
gitdb=4.0.9=pypi_0
gitpython=3.1.27=pypi_0
idna=3.3=pypi_0
importlib-metadata=4.12.0=pypi_0
importlib-resources=5.8.0=pypi_0
ipykernel=6.15.0=pypi_0
ipython=8.4.0=pypi_0
ipython-genutils=0.2.0=pypi_0
ipywidgets=7.7.1=pypi_0
jedi=0.18.1=pypi_0
jinja2=3.1.2=pypi_0
jsonschema=4.6.1=pypi_0
jupyter-client=7.3.4=pypi_0
jupyter-core=4.10.0=pypi_0
jupyterlab-pygments=0.2.2=pypi_0
jupyterlab-widgets=1.1.1=pypi_0
kiwisolver=1.4.3=pypi_0
libcxx=12.0.0=h2f01273_0
libffi=3.3=hb1e8313_2
markupsafe=2.1.1=pypi_0
matplotlib=3.5.2=pypi_0
matplotlib-inline=0.1.3=pypi_0
mistune=0.8.4=pypi_0
nbclient=0.6.6=pypi_0
nbconvert=6.5.0=pypi_0
nbformat=5.4.0=pypi_0
ncurses=6.3=hca72f7f_3
nest-asyncio=1.5.5=pypi_0
notebook=6.4.12=pypi_0
numpy=1.23.0=pypi_0
opencv-python=4.6.0.66=pypi_0
openssl=1.1.1p=hca72f7f_0
packaging=21.3=pypi_0
pandas=1.4.3=pypi_0
pandocfilters=1.5.0=pypi_0
parso=0.8.3=pypi_0
pexpect=4.8.0=pypi_0
pickleshare=0.7.5=pypi_0
pillow=9.2.0=pypi_0
pip=21.2.4=py38hecd8cb5_0
prometheus-client=0.14.1=pypi_0
prompt-toolkit=3.0.30=pypi_0
protobuf=3.20.1=pypi_0
psutil=5.9.1=pypi_0
ptyprocess=0.7.0=pypi_0
pure-eval=0.2.2=pypi_0
pyarrow=8.0.0=pypi_0
pycparser=2.21=pypi_0
pydeck=0.7.1=pypi_0
pygments=2.12.0=pypi_0
pympler=1.0.1=pypi_0
pyparsing=3.0.9=pypi_0
pyrsistent=0.18.1=pypi_0
python=3.8.13=hdfd78df_0
python-dateutil=2.8.2=pypi_0
pytz=2022.1=pypi_0
pytz-deprecation-shim=0.1.0.post0=pypi_0
pyyaml=6.0=pypi_0
pyzmq=23.2.0=pypi_0
readline=8.1.2=hca72f7f_1
requests=2.28.1=pypi_0
rich=12.4.4=pypi_0
scipy=1.8.1=pypi_0
seaborn=0.11.2=pypi_0
semver=2.13.0=pypi_0
send2trash=1.8.0=pypi_0
setuptools=61.2.0=py38hecd8cb5_0
six=1.16.0=pypi_0
smmap=5.0.0=pypi_0
soupsieve=2.3.2.post1=pypi_0
sqlite=3.38.5=h707629a_0
stack-data=0.3.0=pypi_0
streamlit=1.10.0=pypi_0
terminado=0.15.0=pypi_0
tinycss2=1.1.1=pypi_0
tk=8.6.12=h5d9f67b_0
toml=0.10.2=pypi_0
toolz=0.11.2=pypi_0
torch=1.12.0=pypi_0
torchvision=0.13.0=pypi_0
tornado=6.2=pypi_0
tqdm=4.64.0=pypi_0
traitlets=5.3.0=pypi_0
typing-extensions=4.3.0=pypi_0
tzdata=2022.1=pypi_0
tzlocal=4.2=pypi_0
urllib3=1.26.9=pypi_0
validators=0.20.0=pypi_0
wcwidth=0.2.5=pypi_0
webencodings=0.5.1=pypi_0
wheel=0.37.1=pyhd3eb1b0_0
widgetsnbextension=3.6.1=pypi_0
xz=5.2.5=hca72f7f_1
zipp=3.8.0=pypi_0
zlib=1.2.12=h4dc903c_2
